bug fixes after refactoring + tests modified

+++ app/Http/Controllers/ClientsController.php
@@ -6,6 +6,8 @@ use App\Helpers\General;
 use App\Helpers\Sanitize;
 use App\Models\ClientsCustomField;
 use App\Modules\ArrangedDataTableColumns\ArrangedDataTableColumns;
+use App\Modules\ArrangedDataTableColumns\Exceptions\InvalidDataProvidedException;
+use App\Modules\CustomFields\Exceptions\InvalidCustomFieldsException;
 use App\Services\Client\ClientService;
 use App\Services\Client\Exceptions\ClientException;
 use Exception;
@@ -25,8 +27,9 @@ class ClientsController extends Controller{
 
 		try{
 			$this->client_service->save($request);
-		}catch(ClientException $e){
-			return response($e->getMessage(), $e->getCode());
+			return response(['message' => 'Client saved successfully', 'validity' => 'client_saved'], 200);
+		}catch(ClientException|InvalidCustomFieldsException $e){
+			return response(['message' => $e->getMessage(), 'validity' => $e->getValidity()], $e->getCode());
 		}catch(Exception $e){
 			return General::wentWrong();
 		}
@@ -38,7 +41,7 @@ class ClientsController extends Controller{
 		try{
 			return $this->client_service->fetchIndex($request);
 		}catch(ClientException $e){
-			return response($e->getMessage(), $e->getCode());
+			return response(['message' => $e->getMessage(), 'validity' => $e->getValidity()], $e->getCode());
 		}catch(Exception $e){
 			return General::wentWrong();
 		}
@@ -51,18 +54,32 @@ class ClientsController extends Controller{
 	
 
 	public function saveArrangedColumns(Request $request){
-		return $this->arranged_data_table_columns->saveArrangedColumnsData($request, ClientsCustomField::class, 'clients', 'clients', 'client');
+
+		try{
+			$this->arranged_data_table_columns->saveArrangedColumnsData($request, ClientsCustomField::class, 'clients', 'clients', 'client');
+			return response(['message' => 'Saved successfully', 'validity' => 'saved_success'], 200);
+		}catch(InvalidDataProvidedException $e){
+			return response(['message' => $e->getMessage(), 'validity' => $e->getValidity()], $e->getCode());
+		}catch(Exception $e){
+			return General::wentWrong();
+		}
+
 	}
 
 	public function destroy(Request $request){
 		
 		$ids = $request->input('ids');
+
+		if(!$ids){
+			return response(['message' => 'No valid IDs provided', 'validity' => 'invalid_ids'], config('global.error_code'));
+		}
 		
 		try{
 			$ids = Sanitize::recursive($ids);
 			$this->client_service->deleteClients($ids);
+			return response(['message' => 'Client(s) deleted successfully', 'validity' => 'delete_success'], 200);
 		}catch(ClientException $e){
-			return response($e->getMessage(), $e->getCode());
+			return response(['message' => $e->getMessage(), 'validity' => $e->getValidity()], $e->getCode());
 		}catch(Exception $e){
 			return General::wentWrong();
 		}
@@ -78,7 +95,7 @@ class ClientsController extends Controller{
 			return $this->client_service->fetchSingleClientById($id);
 
 		}catch(ClientException $e){
-			return response($e->getMessage(), $e->getCode());
+			return response(['message' => $e->getMessage(), 'validity' => $e->getValidity()], $e->getCode());
 		}catch(Exception $e){
 			return General::wentWrong();
 		}
@@ -88,8 +105,9 @@ class ClientsController extends Controller{
 	public function update(Request $request){
 		try{
 			$this->client_service->update($request);
+			return response(['message' => 'Client updated successfully', 'validity' => 'client_updated'], 200);
 		}catch(ClientException $e){
-			return response($e->getMessage(), $e->getCode());
+			return response(['message' => $e->getMessage(), 'validity' => $e->getValidity()], $e->getCode());
 		}catch(Exception $e){
 			return General::wentWrong();
 		}
